Finalize refactoring and fix block imports
